+++ untitled2/.idea/artifacts/untitled2_war_exploded.xml
@@ -7,6 +7,9 @@
         <element id="directory" name="classes">
           <element id="module-output" name="untitled2" />
         </element>
+        <element id="directory" name="lib">
+          <element id="library" level="project" name="javax.servlet:jstl:1.2" />
+        </element>
       </element>
     </root>
   </artifact>
