+++ untitled2/src/main/java/DAO/Validator/Validator.java
@@ -1,0 +1,4 @@
+package main.java.DAO.Validator;
+
+public class Validator {
+}
