--- untitled2/src/main/java/Servlets/AddUserServlet.java
@@ -1,59 +1,0 @@
-package main.java.Servlets;
-
-
-import main.java.DAO.User;
-
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.util.Map;
-import java.util.concurrent.atomic.AtomicInteger;
-
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicInteger;
-
-@WebServlet(name = "AddUserServlet")
-
-public class AddUserServlet extends HttpServlet {
-    private Map<Integer, User> users;
-    private AtomicInteger id;
-    public void init(HttpServletRequest request, HttpServletResponse response) throws ServletException{
-        final Object users = getServletContext().getAttribute("users");
-
-        if (users == null || !(users instanceof ConcurrentHashMap)) {
-
-            throw new IllegalStateException("You're repo does not initialize!");
-        } else {
-
-            this.users = (ConcurrentHashMap<Integer, User>) users;
-        }
-
-        id = new AtomicInteger(2);
-    }
-    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        request.setCharacterEncoding("UTF-8");
-
-        //if (Utils.requestIsValid(request)) {
-
-            final String name = request.getParameter("name");
-            final String age = request.getParameter("age");
-
-            final User user = new User();
-            final int id = this.id.getAndIncrement();
-            user.setId(id);
-            user.setAge(Integer.valueOf(age));
-            user.setName(name);
-
-            users.put(id, user);
-       // }
-
-        response.sendRedirect(request.getContextPath() + "/");
-    }
-
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-
-    }
-}
