--- untitled2/src/main/java/Servlets/ContextListenerServlet.java
@@ -1,44 +1,0 @@
-package main.java.Servlets;
-
-import main.java.DAO.User;
-import main.java.Validation.Validator;
-
-import javax.servlet.ServletContext;
-import javax.servlet.ServletContextEvent;
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-import javax.servlet.ServletContext;
-import javax.servlet.ServletContextEvent;
-import javax.servlet.ServletContextListener;
-import javax.servlet.annotation.WebListener;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-@WebServlet(name = "ContextListenerServlet")
-public class ContextListenerServlet implements ServletContextListener {
-    private Map<Integer, User> users;
-    @Override
-    public void contextInitialized(ServletContextEvent event){
-        final ServletContext servletContext =
-                 event.getServletContext();
-
-        users = new ConcurrentHashMap<>();
-
-        servletContext.setAttribute("users", users);
-
-        final User user = Validator.createStubUser(1, "Первый", 10);
-        this.users.put(user.getId(), user);
-    }
-
-    @Override
-    public void contextDestroyed(ServletContextEvent servletContextEvent) {
-        users = null;
-    }
-
-
-}
