--- untitled2/src/main/java/Servlets/RemuveUserServlet.java
@@ -1,43 +1,0 @@
-package main.java.Servlets;
-
-import main.java.DAO.User;
-import main.java.Validation.Validator;
-
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
-@WebServlet(name = "RemuveUserServlet")
-public class RemuveUserServlet extends HttpServlet {
-    private Map<Integer, User> users;
-    @Override
-    public void init() throws ServletException {
-
-        final Object users = getServletContext().getAttribute("users");
-
-        if (users == null || !(users instanceof ConcurrentHashMap)) {
-
-            throw new IllegalStateException("You're repo does not initialize!");
-        } else {
-
-            this.users = (ConcurrentHashMap<Integer, User>) users;
-        }
-    }
-    @Override
-    protected void doPost(HttpServletRequest req, HttpServletResponse resp)
-            throws ServletException, IOException {
-
-        req.setCharacterEncoding("UTF-8");
-
-        if (Validator.idIsNumber(req)) {
-            users.remove(Integer.valueOf(req.getParameter("id")));
-        }
-
-        resp.sendRedirect(req.getContextPath() + "/");
-    }
-}
