--- untitled2/src/main/java/Servlets/UpdateUserServlet.java
@@ -1,68 +1,0 @@
-package main.java.Servlets;
-
-import main.java.DAO.User;
-import main.java.Validation.Validator;
-
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
-@WebServlet(name = "UpdateUserServlet")
-public class UpdateUserServlet extends HttpServlet {
-    private Map<Integer, User> users;
-
-    @Override
-    public void init() throws ServletException {
-
-        final Object users = getServletContext().getAttribute("users");
-
-        if (users == null || !(users instanceof ConcurrentHashMap)) {
-
-            throw new IllegalStateException("You're repo does not initialize!");
-        } else {
-
-            this.users = (ConcurrentHashMap<Integer, User>) users;
-        }
-    }
-
-    @Override
-    protected void doPost(HttpServletRequest req, HttpServletResponse resp)
-            throws ServletException, IOException {
-
-        req.setCharacterEncoding("UTF-8");
-
-        final String id = req.getParameter("id");
-        final String name = req.getParameter("name");
-
-        final User user = users.get(Integer.parseInt(id));
-        user.setName(name);
-
-        resp.sendRedirect(req.getContextPath() + "/");
-    }
-
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp)
-            throws ServletException, IOException {
-
-        final String id = req.getParameter("id");
-
-        if (Validator.idIsInvalid(id, users)) {
-            resp.sendRedirect(req.getContextPath() + "/");
-            return;
-        }
-
-        final User user = users.get(Integer.parseInt(id));
-        req.setAttribute("user", user);
-
-        req.getRequestDispatcher("/WEB-INF/view/update.jsp")
-                .forward(req, resp);
-    }
-
-
-
-}
